Refactor startup components to be more streamlined
