+++ client/screens/CoursesPage/index.jsx
