+++ client/screens/EventPage/EventPageView/Attendees/index.jsx
