+++ client/screens/EventPage/EventPageView/CurrentEvent.jsx
