+++ client/screens/EventPage/EventPageView/DeleteEvent.jsx
