+++ client/screens/EventPage/EventPageView/index.jsx
