+++ client/screens/EventPage/index.jsx
